avr: clean up ATtiny definitions
Add definitions for constants like PB0.
