+++ src/machine/board_digispark.go
@@ -8,12 +8,12 @@ func CPUFrequency() uint32 {
 }
 
 const (
-	P0 Pin = 0
-	P1 Pin = 1
-	P2 Pin = 2
-	P3 Pin = 3
-	P4 Pin = 4
-	P5 Pin = 5
+	P0 Pin = PB0
+	P1 Pin = PB1
+	P2 Pin = PB2
+	P3 Pin = PB3
+	P4 Pin = PB4
+	P5 Pin = PB5
 
 	LED = P1
 )
